fix handling of transitive dependencies
This fixes issue with missing dependencies like:
```
> Could not find com.google.protobuf:protobuf-bom:3.10.0.
```
Which was not reported for `com.google.protobuf:protobuf-java:3.10.0`.

This is necessary for Gradle 6+, but for now I'm leaving this disabled.

It's not enough to track `dependencies` field. We also need to scan the
`dependencyManagement` field for transitive dependencies. Most of them
are in `test` scope or without a scope, but some like `import` are needed.

Dependency Management allows to consolidate and centralize the management
of dep. versions without adding deps which are inherited by all children.
Useful when you have a set of projects that inherits a common parent.

For more details see:
- https://maven.apache.org/guides/introduction/introduction-to-dependency-mechanism.html#Dependency_Management

Signed-off-by: Jakub Sokołowski <jakub@status.im>